requirements.txt
alembic==1.0.1
Click==7.0
Flask==1.0.2
Flask-Login==0.4.1
Flask-Migrate==2.3.0
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mysqlclient==1.3.13
python-dateutil==2.7.4
python-dotenv==0.9.1
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.12
Werkzeug==0.14.1
